Move all drawing into paintComponent.
Rather than doing the crazy draw to an image every time something
changes, we will simply redraw everything when paintComponent is
called. Instead of calling updateAll every time something changes,
we'll trigger a repaint which will avoid redrawing our backing
image a bunch if we have multiple changes. If we need to do
intelligent diffing to avoid redrawing everything than we will,
but we'll do it entirely in the Board component and we'll diff
at repaint time rather than trying to have the changes to the
ViewModel itself figure out what's important.